io/kqueue: extract socket-side code into kqueue_socket.c
Commit 4b of the macOS backend series.  Pure code move of the
socket-side kqueue code (io_handler_init/fini/main_loop, accept,
connect, read, write, resubmit, heartbeat primitives, io_send_request,
io_request_read_op, io_request_write_op) from backend_kqueue.c
into a new lib/io/kqueue_socket.c.

After this commit the FreeBSD-specific file-I/O bits (aio_read,
aio_write, io_backend_init/main_loop, handle_backend_{pread,pwrite})
are the only thing left in backend_kqueue.c (361 lines, down from
1255).  The socket code (~900 lines) is now in kqueue_socket.c
and compiled under either IO_BACKEND_KQUEUE (FreeBSD, alongside
backend_kqueue.c's aio) or IO_BACKEND_DARWIN (macOS, alongside
the thread-pool backend_darwin.c from commit 5).

Move discipline: diff should be byte-identical move per
backend-extraction.md §6.  Grep confirms zero drive-by fixes.

Makefile.am: both IO_BACKEND_KQUEUE and IO_BACKEND_DARWIN now
compile kqueue_common.c + kqueue_socket.c.  The former additionally
adds backend_kqueue.c; the latter will add backend_darwin.c in
commit 7.

Linux build unchanged (neither file is compiled when
IO_BACKEND_LIBURING).  FreeBSD runtime unchanged -- same object
code, just now produced from two .c files instead of one.
Smoke re-verify on witchie before this PR merges.

Splits commit 4 per the plan addendum W-5: 4a introduced
kqueue_common.{h,c}; 4b (this commit) introduces kqueue_socket.c.
