+++ lib/io/Makefile.am
@@ -43,10 +43,12 @@ endif
 if IO_BACKEND_KQUEUE
 libreffs_io_la_SOURCES += backend_kqueue.c
 libreffs_io_la_SOURCES += kqueue_common.c
+libreffs_io_la_SOURCES += kqueue_socket.c
 endif
 
 if IO_BACKEND_DARWIN
 libreffs_io_la_SOURCES += kqueue_common.c
-# backend_darwin.c and kqueue_socket.c wired in by later commits
-# (commits 5-7) of the macOS backend series.
+libreffs_io_la_SOURCES += kqueue_socket.c
+# backend_darwin.c wired in by later commits (commits 5-7)
+# of the macOS backend series.
 endif
